requirements.txt
fastapi
uvicorn
supabase
python-dotenv
requests
python-multipart
gpustat
